+++ src/Sulu/Bundle/CoreBundle/Tests/DependencyInjection/SuluCoreExtensionTest.php
@@ -16,6 +16,15 @@ use Sulu\Bundle\CoreBundle\DependencyInjection\SuluCoreExtension;
 
 class SuluCoreExtensionTest extends AbstractExtensionTestCase
 {
+    protected function setUp()
+    {
+        parent::setUp();
+
+        $this->container->setParameter('kernel.bundles', []);
+        $this->container->setParameter('kernel.cache_dir', __DIR__);
+    }
+
+
     protected function getContainerExtensions()
     {
         return [