+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 CHANGELOG for Sulu
 ==================
     
+* dev-master
+    * HOTFIX     #1991   [Rest]             Added metadata for field-descriptors
+    
 * 1.1.9 (2016-02-05)
     * ENHANCEMENT #1978 [SecurityBundle]    Made url for resetting password configurable via static variable
     * HOTFIX      #1976 [MediaBundle]       Moved delete collection to drop-down to avoid misunderstandings